// src/enums/CronRunNames.ts
enum CronRunNames {
  PRODUCT_LEASE_INTEREST_APPLYER = 'PRODUCT_LEASE_INTEREST_APPLYER',
  MULTI_STAGE_PRODUCT_LEASE_INTEREST_APPLYER = 'MULTI_STAGE_PRODUCT_LEASE_INTEREST_APPLYER',
  CLOSED_PAID_ORDER = 'CLOSED_PAID_ORDER',
  QUOTE_REQUEST_EXPIRE = 'QUOTE_REQUEST_EXPIRE',
  CHARGE_BUYER_ON_PAYMENT_DEFAULT = 'CHARGE_BUYER_ON_PAYMENT_DEFAULT',
  NOTIFY_BUYER_BEFORE_CLOSE_ORDER = 'NOTIFY_BUYER_BEFORE_CLOSE_ORDER',
  CONFIRM_WALLET_ORDERS = 'CONFIRM_WALLET_ORDERS',
  C_STORE_ORDER_LEASE_CHARGE = 'C_STORE_ORDER_LEASE_CHARGE',
}

export default CronRunNames